fix(web): inline the multi-channel Text Murph choices in the top-up success dialog
Resolves the final-review Experience Collapse finding: with several
resolved channels the fulfilled dialog no longer opens a second modal;
the channel rows render inline under a Text Murph label with Close as
the secondary action. The stateless rows move to a shared
MurphContactChannelRows fragment; the browser-handoff picker keeps its
own modal shell and behavior.
