+++ apps/web/test/biomarker-design-studies.test.tsx
@@ -121,6 +121,7 @@ test("design page routes the biomarker studies through the dedicated sections ta
   expect(sectionsMarkup).toContain("Choose amount");
   expect(sectionsMarkup).toContain('data-design-state="usage-added-text-murph"');
   expect(sectionsMarkup).toContain("Preview usage added with Text Murph");
+  expect(sectionsMarkup).toContain("Preview usage added with channel choices");
   expect(sectionsMarkup).toContain("personal-usage-credit-owner");
   expect(sectionsMarkup).toContain("Pulse included AI usage");
   expect(sectionsMarkup).toContain("Static owner-layout preview");
